+++ common/src/main/java/fr/loudo/narrativecraft/client/editors/dialog/ClientCharacterDialogEditorMaker.java
@@ -116,7 +116,7 @@ public class ClientCharacterDialogEditorMaker implements EditorMaker {
         }
         minecraft.gui.setScreen(null);
         Services.PACKET.sendToServer(BiStopEditorMaker.INSTANCE);
-        playerSession.getCharactersInWorld().remove(character);
+        playerSession.getCharactersInWorld().values().remove(character);
     }
 
     @Override