+++ common/src/main/java/fr/loudo/narrativecraft/events/client/OnRenderNameTagEvent.java
@@ -27,7 +27,6 @@ import fr.loudo.narrativecraft.client.ClientNarrativeCraftMod;
 import fr.loudo.narrativecraft.client.session.ClientPlayerSession;
 import fr.loudo.narrativecraft.narrative.character.CharacterType;
 import fr.loudo.narrativecraft.narrative.character.ICharacterStory;
-import fr.loudo.narrativecraft.utils.Utils;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.player.Player;
 
@@ -38,13 +37,10 @@ public class OnRenderNameTagEvent {
         if (!playerSession.isInStory()) return true;
         if (!(entity instanceof Player player)) return true;
 
-        for (ICharacterStory characterStory : playerSession.getCharactersInWorld()) {
-            if (characterStory.getId().equals(Utils.resolveCharacterId(player.getGameProfile()))
-                    && characterStory.getCharacterType() == CharacterType.NPC) {
-                return false;
-            }
-        }
+        ICharacterStory characterStory =
+                playerSession.getCharacterByProfileId(player.getGameProfile().id());
+        if (characterStory == null) return true;
 
-        return true;
+        return characterStory.getCharacterType() != CharacterType.NPC;
     }
 }