+++ common/src/main/java/fr/loudo/narrativecraft/utils/FakePlayer.java
@@ -23,10 +23,7 @@
 
 package fr.loudo.narrativecraft.utils;
 
-import com.google.common.collect.ImmutableMultimap;
 import com.mojang.authlib.GameProfile;
-import com.mojang.authlib.properties.Property;
-import com.mojang.authlib.properties.PropertyMap;
 import fr.loudo.narrativecraft.mixin.accessor.AvatarAccessor;
 import fr.loudo.narrativecraft.narrative.character.ICharacterStory;
 import io.netty.channel.ChannelFutureListener;
@@ -54,15 +51,11 @@ import org.jetbrains.annotations.Nullable;
 
 // FakePlayer class from Forge
 public class FakePlayer extends ServerPlayer {
-    public static final String CHARACTER_ID_PROPERTY = "nc_character_id";
     private static final ClientInformation DEFAULT_CLIENT_INFO = ClientInformation.createDefault();
     private final boolean isInvulnerable;
 
     public static GameProfile createCharacterProfile(ICharacterStory characterStory, String name) {
-        PropertyMap properties = new PropertyMap(ImmutableMultimap.of(
-                CHARACTER_ID_PROPERTY,
-                new Property(CHARACTER_ID_PROPERTY, characterStory.getId().toString())));
-        return new GameProfile(UUID.randomUUID(), name, properties);
+        return new GameProfile(UUID.randomUUID(), name);
     }
 
     public FakePlayer(ServerLevel level, GameProfile profile, boolean isInvulnerable) {