+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 - Capes mod overriding main character skin to local player
 - NPC name tag shown in a story
 - Character from camera angle items not correctly applied
+- Occasional and random crash when getting character skin